LPS-174541 Rename package
