+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/upgrade/registry/ObjectServiceUpgradeStepRegistrator.java
@@ -268,7 +268,7 @@ public class ObjectServiceUpgradeStepRegistrator
 
 		registry.register(
 			"4.0.0", "4.1.0",
-			new com.liferay.object.internal.upgrade.v4_0_1.
+			new com.liferay.object.internal.upgrade.v4_1_0.
 				ObjectDefinitionUpgradeProcess());
 	}
